backend/requirements.txt


# Web Framework
flask==2.0.1
gunicorn==20.1.0
flask-cors==3.0.10

# Database
pymongo==4.3.3

# Image Processing
qrcode[pil]==7.3.1
pillow==9.2.0
svglib==1.4.1
reportlab==3.6.12

# Cloud Services
cloudinary==1.30.0

# Utilities
python-dotenv==0.21.0
requests==2.28.1
numpy==1.23.4
pandas==1.5.1

# Image Analysis
opencv-python-headless==4.6.0.66
pyzbar==0.1.9

# API & Data Handling
pyjwt==2.6.0
fake-useragent==0.1.11
requests-toolbelt==0.10.1

# Visualization
matplotlib==3.6.2
seaborn==0.12.1
plotly==5.11.0

# PDF & Barcode Handling
PyPDF2==2.11.1
python-barcode==0.14.0
barcode==1.0.2

# Additional Utilities
beautifulsoup4==4.11.1
lxml==4.9.1
openpyxl==3.0.10
geopy==2.3.0
